Refactoring Work: * Made classes that rely on Tool work with BaseTool * Segregated AbstractTool and CreationTool into legacy classes that   other features may continue to use until they get refactored * Added new, simplified architecture for simpler creation tools (like   TextCreationTool) * Fixed Typo in ClickListeningTool
