+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/tool/ClickListeningTool.java
@@ -2,5 +2,5 @@ package org.jhotdraw.draw.tool;
 
 import java.awt.event.MouseListener;
 
-public interface ClickListeingTool extends MouseListener {
+public interface ClickListeningTool extends MouseListener {
 }
