+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/draw/DrawApplicationModel.java
@@ -34,12 +34,7 @@ import org.jhotdraw.draw.figure.TriangleFigure;
 import org.jhotdraw.draw.decoration.ArrowTip;
 import org.jhotdraw.draw.liner.CurvedLiner;
 import org.jhotdraw.draw.liner.ElbowLiner;
-import org.jhotdraw.draw.tool.BezierTool;
-import org.jhotdraw.draw.tool.ConnectionTool;
-import org.jhotdraw.draw.tool.CreationTool;
-import org.jhotdraw.draw.tool.ImageTool;
-import org.jhotdraw.draw.tool.TextAreaCreationTool;
-import org.jhotdraw.draw.tool.TextCreationTool;
+import org.jhotdraw.draw.tool.*;
 import org.jhotdraw.gui.JFileURIChooser;
 import org.jhotdraw.gui.action.ButtonFactory;
 import org.jhotdraw.util.*;
@@ -124,7 +119,7 @@ public class DrawApplicationModel extends DefaultApplicationModel {
         ButtonFactory.addSelectionToolTo(tb, editor, drawingActions, selectionActions);
         tb.addSeparator();
         AbstractAttributedFigure af;
-        CreationTool ct;
+        AbstractCreationTool ct;
         ConnectionTool cnt;
         ConnectionFigure lc;
         ButtonFactory.addToolTo(tb, editor, new CreationTool(new RectangleFigure()), "edit.createRectangle", labels);
