+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/draw/DrawingPanel.java
@@ -36,11 +36,7 @@ import org.jhotdraw.draw.action.*;
 import org.jhotdraw.draw.decoration.ArrowTip;
 import org.jhotdraw.draw.liner.CurvedLiner;
 import org.jhotdraw.draw.liner.ElbowLiner;
-import org.jhotdraw.draw.tool.BezierTool;
-import org.jhotdraw.draw.tool.ConnectionTool;
-import org.jhotdraw.draw.tool.CreationTool;
-import org.jhotdraw.draw.tool.TextAreaCreationTool;
-import org.jhotdraw.draw.tool.TextCreationTool;
+import org.jhotdraw.draw.tool.*;
 import org.jhotdraw.gui.JPopupButton;
 import org.jhotdraw.gui.action.ButtonFactory;
 import org.jhotdraw.undo.UndoRedoManager;
@@ -192,7 +188,7 @@ public class DrawingPanel extends JPanel {
         ButtonFactory.addSelectionToolTo(tb, editor, drawingActions, selectionActions);
         tb.addSeparator();
         AbstractAttributedFigure af;
-        CreationTool ct;
+        AbstractCreationTool ct;
         ConnectionTool cnt;
         ConnectionFigure lc;
         ButtonFactory.addToolTo(tb, editor, new CreationTool(new RectangleFigure()), "edit.createRectangle", labels);
