+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/gui/ToolsToolBar.java
@@ -20,6 +20,7 @@ import static org.jhotdraw.draw.AttributeKeys.PATH_CLOSED;
 import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.draw.DrawingView;
 import org.jhotdraw.draw.action.*;
+import org.jhotdraw.draw.tool.AbstractCreationTool;
 import org.jhotdraw.draw.tool.CreationTool;
 import org.jhotdraw.draw.tool.TextAreaCreationTool;
 import org.jhotdraw.draw.tool.TextCreationTool;
@@ -74,7 +75,7 @@ public class ToolsToolBar extends AbstractToolBar {
                 p.setLayout(layout);
                 GridBagConstraints gbc;
                 AbstractButton btn;
-                CreationTool creationTool;
+                AbstractCreationTool abstractCreationTool;
                 PathTool pathTool;
                 TextCreationTool textTool;
                 TextAreaCreationTool textAreaTool;
@@ -91,16 +92,16 @@ public class ToolsToolBar extends AbstractToolBar {
                 p.add(btn, gbc);
                 labels.configureToolBarButton(btn, "selectionTool");
                 attributes = new HashMap<AttributeKey<?>, Object>();
-                btn = ButtonFactory.addToolTo(this, editor, creationTool = new CreationTool(new SVGRectFigure(), attributes), "createRectangle", labels);
-                creationTool.setToolDoneAfterCreation(false);
+                btn = ButtonFactory.addToolTo(this, editor, abstractCreationTool = new CreationTool(new SVGRectFigure(), attributes), "createRectangle", labels);
+                abstractCreationTool.setToolDoneAfterCreation(false);
                 btn.setUI((PaletteButtonUI) PaletteButtonUI.createUI(btn));
                 gbc = new GridBagConstraints();
                 gbc.gridx = 0;
                 gbc.gridy = 1;
                 gbc.insets = new Insets(3, 0, 0, 0);
                 p.add(btn, gbc);
-                btn = ButtonFactory.addToolTo(this, editor, creationTool = new CreationTool(new SVGEllipseFigure(), attributes), "createEllipse", labels);
-                creationTool.setToolDoneAfterCreation(false);
+                btn = ButtonFactory.addToolTo(this, editor, abstractCreationTool = new CreationTool(new SVGEllipseFigure(), attributes), "createEllipse", labels);
+                abstractCreationTool.setToolDoneAfterCreation(false);
                 btn.setUI((PaletteButtonUI) PaletteButtonUI.createUI(btn));
                 gbc = new GridBagConstraints();
                 gbc.gridx = 1;
@@ -118,8 +119,8 @@ public class ToolsToolBar extends AbstractToolBar {
                 attributes = new HashMap<AttributeKey<?>, Object>();
                 attributes.put(AttributeKeys.FILL_COLOR, null);
                 attributes.put(PATH_CLOSED, false);
-                btn = ButtonFactory.addToolTo(this, editor, creationTool = new CreationTool(new SVGPathFigure(), attributes), "createLine", labels);
-                creationTool.setToolDoneAfterCreation(false);
+                btn = ButtonFactory.addToolTo(this, editor, abstractCreationTool = new CreationTool(new SVGPathFigure(), attributes), "createLine", labels);
+                abstractCreationTool.setToolDoneAfterCreation(false);
                 btn.setUI((PaletteButtonUI) PaletteButtonUI.createUI(btn));
                 gbc = new GridBagConstraints();
                 gbc.gridx = 1;
